pyproject.toml
[tool.poetry]
name = "IPAMviewer"
version = "0.4.0"
description = "Convenient IPAM's subnet viewer from phpIPAM for kiosk display"
authors = ["Anemys <abdelmalik.kerbadou@edu.esiee.fr>", "Redowla <camille.andias@gmail.com>", "Blackjack <theo.peresse-gourbil@edu.esiee.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
flask = "^2.2.2"
tzdata = "^2022.7"
scp = "^0.14.5"
paramiko = "^3.0.0"
python-dotenv = "^0.21.1"

[tool.poetry.scripts]
ipamviewer = "ipamviewer.app:main"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
